change ShareNavigation to shareNavigation and move to navigation folder
